Refactor constrains processor, extract assertion logic into trait
